+++ framework/packages/standards/dex/src/msg.rs
@@ -5,13 +5,13 @@ use abstract_std::{
     adapter,
     objects::{
         fee::{Fee, UsageFee},
-        pool_id::UncheckedPoolAddress,
+        pool_id::{PoolAddressBase, UncheckedPoolAddress},
         AnsAsset, AssetEntry, DexAssetPairing,
     },
     AbstractError, AbstractResult,
 };
 use cosmwasm_schema::QueryResponses;
-use cosmwasm_std::{Addr, CosmosMsg, Decimal, Uint128};
+use cosmwasm_std::{Addr, Api, CosmosMsg, Decimal, Uint128};
 use cw_asset::{AssetBase, AssetInfoBase};
 
 pub use crate::action::DexAction;
@@ -93,6 +93,40 @@ pub enum DexExecuteMsg {
         /// The action to perform
         action: DexAction,
     },
+}
+
+#[cosmwasm_schema::cw_serde]
+/// Swap node for swap route
+pub struct SwapNode<T: cw_address_like::AddressLike> {
+    /// Pool id of the swap
+    pub pool_id: PoolAddressBase<T>,
+    /// Asset in return from the swap
+    pub ask_asset: AssetInfoBase<T>,
+}
+
+impl SwapNode<String> {
+    /// Validate data contained in an _unchecked_ **swap node** instance; return a new _checked_
+    /// **swap node** instance:
+    /// * For Contract addresses, assert its address is valid
+    ///
+    ///
+    /// ```rust,no_run
+    /// use cosmwasm_std::{Addr, Api};
+    /// use abstract_std::{abstract_dex_standard::msg::SwapNode, AbstractResult};
+    ///
+    /// fn validate_swap_node(api: &dyn Api, swap_node_unchecked: &SwapNode<String>) {
+    ///     match swap_node_unchecked.check(api) {
+    ///         Ok(info) => println!("swap node is valid: {}", info.to_string()),
+    ///         Err(err) => println!("swap node is invalid! reason: {}", err),
+    ///     }
+    /// }
+    /// ```
+    pub fn check(self, api: &dyn Api) -> AbstractResult<SwapNode<Addr>> {
+        Ok(SwapNode {
+            pool_id: self.pool_id.check(api)?,
+            ask_asset: self.ask_asset.check(api, None)?,
+        })
+    }
 }
 
 /// Query messages for the dex adapter